api/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
reportlab
fpdf
PyPDF2
python-docx
markdown
pymupdf
gunicorn